% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pformat.R
\name{is_integer}
\alias{is_integer}
\title{Tells if a string represents an integer}
\usage{
is_integer(s)
}
\arguments{
\item{s}{a string}
}
\value{
a logical vector
}
\description{
Tells if a string represents an integer
}
\details{
Doesn't work if the string ends with an L
}

