% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.R
\name{pformatter}
\alias{pformatter}
\title{Formats an object for pretty printing}
\usage{
pformatter(value, format_spec)
}
\arguments{
\item{value}{any R object (conceptually)}

\item{format_spec}{a character vector containing a specification}
}
\value{
the formatted string representing the value
}
\description{
Format an object according to a format specification
}
\details{
\code{pformatter} is a generic function
}

