% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pformat.R
\name{ppformat}
\alias{ppformat}
\title{Prints formatted output}
\usage{
ppformat(format_string, ..., with = NULL)
}
\arguments{
\item{format_string}{a format string (See details for specification) or a
\code{pformat.compiled} object returned from \code{pformat_parse()}}
}
\description{
Just like pformat(), but prints to the standard output
}
