% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_integer}
\alias{is_integer}
\title{Tells if a string represents an integer}
\usage{
is_integer(v)
}
\arguments{
\item{v}{a string vector}
}
\value{
a logical vector
}
\description{
Checks if the characters in a string are in the interval '0' .. '9'
}

