% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pformat.R
\name{pformat}
\alias{pformat}
\title{Perfoms formatted printing}
\usage{
pformat(format_string, ...)
}
\arguments{
\item{format_string}{a format string (See details for specification) or a
\code{pformat.compiled} object returned from \code{pformat_parse()}}

\item{with}{a \code{(pair)list}, \code{data.frame} or \code{environment} on 
which named fields or expressions can be evaluated}
}
\value{
returns a `character` vector with the output
}
\description{
provides pretty string formatting capabilities through
a format specification
}
\details{
Numbered fields start from 1 (not from 0, as in Python.)

Since internally \code{pformat()} uses \code{paste0()}, it is naturally 
vectorized. 

Non-named fields (manually or automatically numbered) are evaluated only over
the list of arguments passed. You can't write expressions involving non-named
fields. In Python, you can write \code{"{0.real}".format(1+2j)}
and it will return a string with the real part of the number. You can't do
this here, it's a design choice. Allowing it would imply limiting the 
kind of expressions one could use as a field, and expressions surely provide
more power than numbered fields.

On named fields (and expressions in general), names are searched first on the 
keyword arguments, then on \code{names(with)} and only then on the caller's
environment.
}

